pyproject.toml
[tool.poetry]
name = "dpy_voice_chart"
version = "0.1.0"
description = ""
authors = ["Ben Simms <ben@bensimms.moe>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"
matplotlib = "^3.3.4"

[tool.poetry.dev-dependencies]
"discord.py-stubs" = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
